% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_2PLCFA.R
\name{bs_2PLCFA}
\alias{bs_2PLCFA}
\title{Bootstrap Version of GVEM Confirmatory Analysis for M2PL}
\usage{
bs_2PLCFA(gvem_result, boots = 5)
}
\arguments{
\item{gvem_result}{a list that includes exploratory or confirmatory GVEM
results for M2PL models.}

\item{boots}{the number of bootstrap samples; default is 5}
}
\value{
a list containing the following objects:
  \item{boots_a}{item discrimination parameters corrected by bootstrap sampling, a \eqn{J \times K} \code{matrix}}
  \item{boots_b}{item difficulty parameters corrected by bootstrap sampling, a vector of length \eqn{J}}
  \item{sd_a}{stardard errors of item discrimination parameters, a \eqn{J \times K} matrix}
  \item{sd_b}{stardard errors of item difficulty parameters, a vector of length \eqn{J}}
}
\description{
A bootstrap version of GVEM (i.e., GVEM-BS) can be implemented
to correct the bias on item parameters and compute standard errors under M2PL models
}
\examples{
\dontrun{
gvem_result <- gvem_2PLCFA(exampleData_2pl, exampleIndic_cfa2pl)
bs_2PLCFA(gvem_result, boots=10)
}
}
\seealso{
\code{\link{gvem_2PLCFA}},\code{\link{importanceSampling}}
}
