% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C2PL_data}
\alias{C2PL_data}
\title{Simulated Data Set for Confirmatory M2PL Analysis}
\format{
A list of components of the data set:
\tabular{ll}{
\code{ data}\tab Item responses\cr
\tab\cr
\code{model}\tab Loading indicators\cr
\tab\cr
\code{params}\tab True parameters used for generating the item responses
}
}
\usage{
C2PL_data
}
\description{
Responses are simulated based on an M2PL model with 2 factors. The true factor correlations are set as 0.8.
}
\author{
Weicong Lyu <wlyu4@uw.edu>
}
\keyword{datasets}
