% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C3PL_data}
\alias{C3PL_data}
\title{Simulated Data Set for Confirmatory M3PL Analysis}
\format{
A list of components of the data set:
\tabular{ll}{
\code{ data}\tab Item responses\cr
\tab\cr
\code{model}\tab Loading indicators\cr
\tab\cr
\code{params}\tab True parameters used for generating the item responses
}
}
\usage{
C3PL_data
}
\description{
Responses are simulated based on an M3PL model with 2 factors. The true factor correlations are set as 0.8.
}
\author{
Weicong Lyu <weiconglyu@um.edu.mo>
}
\keyword{datasets}
