% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{coef.vemirt_DIF_summary}
\alias{coef.vemirt_DIF_summary}
\title{Extract DIF Items}
\usage{
coef(object)
}
\arguments{
\item{object}{An object of class \code{vemirt_DIF_summary}}
}
\description{
Extract DIF Items
}
\seealso{
\code{\link{summary.vemirt_DIF}}, \code{\link{print.vemirt_DIF_summary}}
}
\author{
Weicong Lyu <weiconglyu@um.edu.mo>
}
