% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1PL_em.R
\name{D1PL_em}
\alias{D1PL_em}
\title{EM Algorithms for DIF Detection in M1PL Models}
\usage{
D1PL_em(
  data,
  model = matrix(1, ncol(data)),
  group = rep(1, nrow(data)),
  a = 1,
  method = "EMM",
  Lambda0 = if (length(unique(group)) == 1) 0 else seq(0.1, 1, by = 0.1),
  level = 10,
  criterion = "BIC",
  iter = 200,
  eps = 0.001,
  c = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{An \eqn{N\times J} binary matrix of item responses (missing responses should be coded as \code{NA})}

\item{model}{A \eqn{J\times K} binary matrix of loading indicators  (all items load on the only dimension by default)}

\item{group}{An \eqn{N} dimensional vector of group indicators from \code{1} to \code{G} (all respondents are in the same group by default)}

\item{a}{A scalar indicating the common discrimination parameter for all the dimensions of all the items (takes \code{1} by default)}

\item{method}{Estimation algorithm, one of \code{'EM'} or \code{'EMM'}}

\item{Lambda0}{A vector of \code{lambda0} values for \eqn{L_1} penalty (\code{lambda} equals \code{sqrt(N) * lambda0})}

\item{level}{Accuracy level, either a number for \code{mvQuad} or a vector indicating the grid for each latent dimension}

\item{criterion}{Information criterion for model selection, one of \code{'BIC'} (recommended), \code{'AIC'}, or \code{'GIC'}}

\item{iter}{Maximum number of iterations}

\item{eps}{Termination criterion on numerical accuracy}

\item{c}{Constant for computing GIC}

\item{verbose}{Whether to show the progress}
}
\value{
An object of class \code{vemirt_DIF}, which is a list containing the following elements:
  \item{N}{Number of respondents}
  \item{niter0}{Number(s) of iterations for initialization}
  \item{fit}{The best (with lowest information criterion) model, which is an element of \code{all}}
  \item{best}{The index of \code{fit} in \code{all}}
  \item{all}{A list of models which has the same length as \code{Lambda0}:}
  \item{ ...$lambda0}{Corresponding element in \code{Lambda0}}
  \item{ ...$lambda}{\code{sqrt(N) * lambda0}}
  \item{ ...$niter}{Number(s) of iterations}
  \item{ ...$Sigma}{Group-level covariance matrices}
  \item{ ...$Mu}{Group-level mean vectors}
  \item{ ...$a}{Slopes for group 1}
  \item{ ...$b}{Intercepts for group 1}
  \item{ ...$gamma}{D1PL parameters for the slopes (all elements are zero)}
  \item{ ...$beta}{D1PL parameters for the intercepts}
  \item{ ...$ll}{Log-likelihood}
  \item{ ...$l0}{Number of nonzero D1PL parameters in \code{gamma} and \code{beta}}
  \item{ ...$AIC}{Akaike Information Criterion: \code{-2*ll+l0*2}}
  \item{ ...$BIC}{Bayesian Information Criterion: \code{-2*ll+l0*log(N)}}
  \item{ ...$GIC}{Generalized Information Criterion: \code{-2*ll+c*l0*log(N)*log(log(N))}}
}
\description{
EM Algorithms for DIF Detection in M1PL Models
}
\examples{
\dontrun{
with(D1PL_data, D1PL_em(data, model, group))}
}
\seealso{
\code{\link{D1PL_gvem}}, \code{\link{coef.vemirt_DIF}}, \code{\link{print.vemirt_DIF}}, \code{\link{summary.vemirt_DIF}}
}
\author{
Weicong Lyu <weiconglyu@um.edu.mo>
}
