% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1PL_gvem.R
\name{D1PL_gvem}
\alias{D1PL_gvem}
\title{GVEM Algorithms for DIF Detection in M1PL Models}
\usage{
D1PL_gvem(
  data,
  model = matrix(1, ncol(data)),
  group = rep(1, nrow(data)),
  method = "IWGVEMM",
  Lambda0 = if (length(unique(group)) == 1) 0 else seq(0.1, 1, by = 0.1),
  criterion = "GIC",
  iter = 200,
  eps = 0.001,
  c = 1,
  S = 10,
  M = 10,
  lr = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{data}{An \eqn{N\times J} binary matrix of item responses (missing responses should be coded as \code{NA})}

\item{model}{A \eqn{J\times K} binary matrix of loading indicators  (all items load on the only dimension by default)}

\item{group}{An \eqn{N} dimensional vector of group indicators from \code{1} to \code{G} (all respondents are in the same group by default)}

\item{method}{Estimation algorithm, one of \code{'GVEM'} or \code{'IWGVEMM'}}

\item{Lambda0}{A vector of \code{lambda0} values for \eqn{L_1} penalty (\code{lambda} equals \code{sqrt(N) * lambda0})}

\item{criterion}{Information criterion for model selection, one of \code{'GIC'} (recommended), \code{'BIC'}, or \code{'AIC'}}

\item{iter}{Maximum number of iterations}

\item{eps}{Termination criterion on numerical accuracy}

\item{c}{Constant for computing GIC}

\item{S}{Sample size for approximating the expected lower bound (\code{'IWGVEMM'} only)}

\item{M}{Sample size for approximating a tighter lower bound (\code{'IWGVEMM'} only)}

\item{lr}{Learning rate for the Adam optimizer (\code{'IWGVEMM'} only)}

\item{verbose}{Whether to show the progress}
}
\value{
An object of class \code{vemirt_DIF}, which is a list containing the following elements:
  \item{N}{Number of respondents}
  \item{niter0}{Number(s) of iterations for initialization}
  \item{fit}{The best (with lowest information criterion) model, which is an element of \code{all}}
  \item{best}{The index of \code{fit} in \code{all}}
  \item{all}{A list of models which has the same length as \code{Lambda0}:}
  \item{ ...$lambda0}{Corresponding element in \code{Lambda0}}
  \item{ ...$lambda}{\code{sqrt(N) * lambda0}}
  \item{ ...$niter}{Number(s) of iterations}
  \item{ ...$SIGMA}{Person-level posterior covariance matrices}
  \item{ ...$MU}{Person-level posterior mean vectors}
  \item{ ...$Sigma}{Group-level covariance matrices}
  \item{ ...$Mu}{Group-level mean vectors}
  \item{ ...$a}{Slopes (fixed to 1)}
  \item{ ...$b}{Intercepts for group 1}
  \item{ ...$beta}{D1PL parameters for the intercepts}
  \item{ ...$RMSE}{Root mean square error of fitted probability of each item for each group}
  \item{ ...$ll}{Estimated lower bound of log-likelihood}
  \item{ ...$l0}{Number of nonzero D1PL parameters in \code{beta}}
  \item{ ...$AIC}{Akaike Information Criterion: \code{-2*ll+l0*2}}
  \item{ ...$BIC}{Bayesian Information Criterion: \code{-2*ll+l0*log(N)}}
  \item{ ...$GIC}{Generalized Information Criterion: \code{-2*ll+c*l0*log(N)*log(log(N))}}
}
\description{
GVEM Algorithms for DIF Detection in M1PL Models
}
\examples{
\dontrun{
with(D1PL_data, D1PL_gvem(data, model, group))}
}
\seealso{
\code{\link{D1PL_pair_em}}, \code{\link{D1PL_em}}, \code{\link{coef.vemirt_DIF}}, \code{\link{print.vemirt_DIF}}, \code{\link{summary.vemirt_DIF}}
}
\author{
Weicong Lyu <weiconglyu@um.edu.mo>
}
