% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2PL_lrt.R
\name{D2PL_lrt}
\alias{D2PL_lrt}
\title{Likelihood Ratio Test for DIF Detection in M2PL Models}
\usage{
D2PL_lrt(data, model, group, unif = F)
}
\arguments{
\item{data}{An \eqn{N\times J} binary matrix of item responses}

\item{model}{A \eqn{J\times K} binary matrix of loading indicators}

\item{group}{An \eqn{N} dimensional vector of group indicators (integers from \code{1} to \code{G})}

\item{unif}{Whether to detect uniform D2PL only}
}
\value{
A list:
  \item{Sigma}{Group-level posterior covariance matrices}
  \item{Mu}{Group-level posterior mean vectors}
  \item{a}{Slopes for group 1}
  \item{b}{Intercepts for group 1}
  \item{gamma}{D2PL parameters for the slopes}
  \item{beta}{D2PL parameters for the intercepts}
}
\description{
Likelihood Ratio Test for DIF Detection in M2PL Models
}
\examples{
\dontrun{
with(D2PL_data, D2PL_lrt(data, model, group))}
}
\seealso{
\code{\link{D2PL_em}}, \code{\link{D2PL_gvem}}
}
\author{
Ruoyi Zhu <zhux0445@uw.edu>
}
