% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_lrt.R
\name{DIF_lrt}
\alias{DIF_lrt}
\title{Likelihood Ratio Test for DIF Detection in 2PL Models}
\usage{
DIF_lrt(data, model, group, unif = F)
}
\arguments{
\item{data}{An \eqn{N\times J} binary matrix of item responses}

\item{model}{A \eqn{J\times K} binary matrix of loading indicators}

\item{group}{An \eqn{N} dimensional vector of group indicators (integers from \code{1} to \code{G})}

\item{unif}{Whether to detect uniform DIF only}
}
\value{
A list:
  \item{Sigma}{Group-level posterior covariance matrices}
  \item{Mu}{Group-level posterior mean vectors}
  \item{a}{Slopes for group 1}
  \item{b}{Intercepts for group 1}
  \item{gamma}{DIF parameters for the slopes}
  \item{beta}{DIF parameters for the intercepts}
}
\description{
Likelihood Ratio Test for DIF Detection in 2PL Models
}
\examples{
\dontrun{
with(exampleDIF, DIF_lrt(data, model, group))}
}
\seealso{
\code{\link{DIF_em}}, \code{\link{DIF_gvem}}
}
\author{
Ruoyi Zhu <zhux0445@uw.edu>
}
