% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{E2PL_data_C1}
\alias{E2PL_data_C1}
\title{Simulated Data Set for Exploratory M2PL Analysis Under C1 Constraints}
\format{
A list of components of the data set:
\tabular{ll}{
\code{ data}\tab Item responses\cr
\tab\cr
\code{model}\tab Loading indicators for (adaptive) lasso penalty\cr
\tab\cr
\code{constrain}\tab Constraint for model identification (\code{'C1'})\cr
\tab\cr
\code{non_pen}\tab Index of an item that is associated with all the factors (\code{NULL} under C1)
}
}
\usage{
E2PL_data_C1
}
\description{
Responses are simulated based on a between-item M2PL model with 5 factors. The true factor correlations are set as 0.1.
}
\keyword{datasets}
