% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{E3PL_data_C1}
\alias{E3PL_data_C1}
\title{Simulated Data Set for Exploratory M3PL Analysis Under C1 Constraint}
\format{
A list of components of the data set:
\tabular{ll}{
\code{ data}\tab Item responses\cr
\tab\cr
\code{model}\tab Loading indicators for (adaptive) lasso penalty\cr
\tab\cr
\code{constrain}\tab Constraint for model identification (\code{'C1'})\cr
\tab\cr
\code{non_pen}\tab Index of an item that is associated with all the factors (\code{NULL} under \code{C1})\cr
\tab\cr
\code{params}\tab True parameters used for generating the item responses
}
}
\usage{
E3PL_data_C1
}
\description{
Responses are simulated based on an M3PL model with 3 factors. The true factor correlations are set as 0.5.
}
\author{
Weicong Lyu <wlyu4@uw.edu>
}
\keyword{datasets}
