% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_emm.R
\name{em_DIF}
\alias{em_DIF}
\title{EM Algorithms for DIF Detection in 2PL Models}
\usage{
em_DIF(
  Y,
  D,
  X,
  method = "EMM",
  unif = F,
  Lambda0 = seq(0.2, 0.7, by = 0.1),
  criterion = "BIC",
  iter = 1000,
  eps = 0.001,
  c = 0.7,
  eta = 1
)
}
\arguments{
\item{Y}{An \eqn{N\times J} binary matrix of item responses}

\item{D}{A \eqn{J\times K} binary matrix of loading indicators}

\item{X}{An \eqn{N} dimensional vector of group indicators (integers from \code{1} to \code{G})}

\item{method}{Estimation algorithm, one of \code{'EM'}, \code{'EMM'} and \code{'Adapt'}}

\item{unif}{Whether to detect uniform DIF only}

\item{Lambda0}{A vector of \code{lambda0} values for \eqn{L_1} penalty (\code{lambda} is \code{sqrt(N) * lambda0})}

\item{criterion}{Information criterion for model selection, one of \code{'BIC'} (recommended), \code{'AIC'} and \code{'GIC'}}

\item{iter}{Maximum number of iterations}

\item{eps}{Termination criterion on numerical accuracy}

\item{c}{Constant for computing GIC}

\item{eta}{Tuning constant for adaptive lasso (\code{'Adapt'} only)}
}
\value{
An object of class \code{vemirt_DIF}, which is a list containing three elements:
  \item{fit}{The best (with lowest information criterion) model, which is an element of \code{all}}
  \item{best}{The location of \code{fit} in \code{all}}
  \item{all}{A list of models whose length is equal to \code{Lambda0}:}
  \item{ ...$lambda0}{Corresponding element in \code{Lambda0}}
  \item{ ...$lambda}{\code{sqrt(N) * lambda0}}
  \item{ ...$iter}{Number(s) of iterations}
  \item{ ...$Sigma}{Group-level posterior covariance matrices}
  \item{ ...$Mu}{Group-level posterior mean vectors}
  \item{ ...$a}{Slopes for group 1}
  \item{ ...$b}{Intercepts for group 1}
  \item{ ...$gamma}{DIF parameters for the slopes}
  \item{ ...$beta}{DIF parameters for the intercepts}
  \item{ ...$ll}{Log-likelihood}
  \item{ ...$l0}{Number of nonzero parameters in \code{gamma} and \code{beta}}
  \item{ ...$AIC}{Akaike Information Criterion}
  \item{ ...$BIC}{Bayesian Information Criterion}
  \item{ ...$GIC}{Generalized Information Criterion}
}
\description{
EM Algorithms for DIF Detection in 2PL Models
}
\examples{
\dontrun{
with(exampleDIF, em_DIF(Y, D, X))}
}
\seealso{
\code{\link{gvemm_DIF}}, \code{\link{lrt_DIF}}, \code{\link{coef.vemirt_DIF}}, \code{\link{print.vemirt_DIF}}
}
