% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData_2pl}
\alias{exampleData_2pl}
\title{Response data set for M2PL}
\format{
A data frame with 2000 respondents and 75 items
}
\usage{
exampleData_2pl
}
\description{
The response data set is simulated based on a
between-item M2PL model with 5 factors. The true factor correlations are set
as 0.1.
}
\keyword{datasets}
