% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData_3pl}
\alias{exampleData_3pl}
\title{Response data set for M3PL}
\format{
A data frame with 2000 respondents and 45 items
}
\usage{
exampleData_3pl
}
\description{
The response data set is simulated based on a
within-item M3PL model with 3 factors. The true factor correlations are set
as 0.1.
}
\keyword{datasets}
