% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleIndic_cfa2pl}
\alias{exampleIndic_cfa2pl}
\title{Factor-loading indicator matrix for M2PL-CFA}
\format{
A data frame with 75 items and 5 factors
}
\usage{
exampleIndic_cfa2pl
}
\description{
The factor-loading indicator matrix can be used as an input for
confirmatory factor analysis.
}
\keyword{datasets}
