% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleIndic_cfa3pl}
\alias{exampleIndic_cfa3pl}
\title{Factor-loading indicator matrix for M3PL-CFA}
\format{
A data frame with 45 items and 3 factors
}
\usage{
exampleIndic_cfa3pl
}
\description{
The factor-loading indicator matrix can be used as an input for
confirmatory factor analysis.
}
\keyword{datasets}
