% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleIndic_efa2pl_c2}
\alias{exampleIndic_efa2pl_c2}
\title{Factor-loading indicator matrix for M2PL-EFA with lasso/ adaptive penalty under constraint 2}
\format{
A data frame with 75 items and 5 factors. Items 1, 16, 31, 46 and 61 can
be combined as a triangular matrix to satisfy constraint 2
}
\usage{
exampleIndic_efa2pl_c2
}
\description{
The factor-loading indicator matrix can be used as an input for
exploratory factor analysis with lasso/ adaptive lasso penalty for constraint 1.
}
\keyword{datasets}
