% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleIndic_efa3pl_c1}
\alias{exampleIndic_efa3pl_c1}
\title{Factor-loading indicator matrix for M3PL-EFA with lasso/ adaptive penalty under constraint 1}
\format{
A data frame with 45 items and 3 factors. Items 1, 16, and 19 can
be combined as an identity matrix to satisfy constraint 1
}
\usage{
exampleIndic_efa3pl_c1
}
\description{
The factor-loading indicator matrix can be used as an input for
exploratory factor analysis with lasso/ adaptive lasso penalty under constraint 1.
}
\keyword{datasets}
