% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleItem_2pl}
\alias{exampleItem_2pl}
\title{True item parameters for M2PL}
\format{
An object of class \code{data.frame} with 75 rows and 6 columns.
}
\usage{
exampleItem_2pl
}
\description{
True item parameters for M2PL
}
\keyword{datasets}
