% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleItem_3pl}
\alias{exampleItem_3pl}
\title{True item parameters for M3PL}
\format{
An object of class \code{data.frame} with 45 rows and 5 columns.
}
\usage{
exampleItem_3pl
}
\description{
True item parameters for M3PL
}
\keyword{datasets}
