% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvem_2PLCFA.R
\name{gvem_2PLCFA}
\alias{gvem_2PLCFA}
\title{Confirmatory M2PL Analysis}
\usage{
gvem_2PLCFA(u, indic, max.iter = 5000, SE.est = FALSE)
}
\arguments{
\item{u}{a \eqn{N \times J} \code{matrix} or a \code{data.frame} that
consists of binary responses of \eqn{N} individuals to \eqn{J} items. The
missing values are coded as \code{NA}}

\item{indic}{a \eqn{J \times K} \code{matrix} or a \code{data.frame} that
describes the factor loading structure of \eqn{J} items to \eqn{K} factors. It
consists of binary values where 0 refers to the item is irrelevant with this factor,
1 otherwise}

\item{max.iter}{the maximum number of iterations for the EM cycle; default is 5000}

\item{SE.est}{whether to estimate SE for item parameters using the updated
supplemented expectation maximization (USEM); default is FALSE}
}
\value{
a list containing the following objects:
  \item{ra}{item discrimination parameters, a \eqn{J \times K} \code{matrix}}
  \item{rb}{item difficulty parameters, vector of length \eqn{J}}
  \item{reta}{variational parameters \eqn{\eta(\xi)}, a \eqn{N \times J} matrix}
  \item{reps}{variational parameters \eqn{\xi}, a \eqn{N \times J} matrix}
  \item{rsigma}{population variance-covariance matrix, a \eqn{K \times K} matrix}
  \item{mu_i}{mean parameter for each person, a \eqn{K \times N} matrix}
  \item{sig_i}{covariance matrix for each person, a \eqn{K \times K \times N} array}
  \item{n}{the number of iterations for the EM cycle}
  \item{Q_mat}{factor loading structure, a \eqn{J \times K} matrix}
  \item{GIC}{model fit index}
  \item{AIC}{model fit index}
  \item{BIC}{model fit index}
  \item{SE}{Standard errors of item parameters, a \eqn{J \times (K+1)} matrix where the last column includes SE estimates for item difficulty parameters}
}
\description{
Confirmatory M2PL Analysis
}
\examples{
\dontrun{
gvem_2PLCFA(exampleData_2pl, exampleIndic_cfa2pl)}
}
\seealso{
\code{\link{sgvem_3PLCFA}},\code{\link{importanceSampling}},\code{\link{bs_2PLCFA}}
}
