% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvem_2PLEFA_adaptlasso.R
\name{gvem_2PLEFA_adaptlasso}
\alias{gvem_2PLEFA_adaptlasso}
\title{Exploratory M2PL Analysis with Adaptive Lasso Penalty}
\usage{
gvem_2PLEFA_adaptlasso(
  u,
  indic,
  max.iter = 5000,
  constrain = "C1",
  non_pen = NULL,
  gamma = 2
)
}
\arguments{
\item{u}{a \eqn{N \times J} \code{matrix} or a \code{data.frame} that
consists of binary responses of \eqn{N} individuals to \eqn{J} items. The
missing values are coded as \code{NA}}

\item{indic}{a \eqn{J \times K} \code{matrix} or a \code{data.frame} that
describes the factor loading structure of \eqn{J} items to \eqn{K} factors. It
consists of binary values where 0 refers to the item is irrelevant with this factor,
1 otherwise. For exploratory factor analysis with adaptive lasso penalty, \code{indic} should be
imposed certain constraints on the a \eqn{K \times K} sub-matrix to ensure identifiability.
The remaining parts do not assume any pre-specified zero structure but instead, the
appropriate lasso penalty would recover the true zero structure. Also see \code{constrain}}

\item{max.iter}{the maximum number of iterations for the EM cycle; default is 5000}

\item{constrain}{the constraint setting: \code{"C1"} or \code{"C2"}. To ensure
identifiability, \code{"C1"} sets a \eqn{K \times K} sub-matrix of \code{indic} to be an
identity matrix, and the variance of the population variance-covariance matrix to be 1.
Instead of setting all off-diagonals of a sub-matrix to be zero, \code{"C2"} keeps the
sub-matrix to be a triangular matrix with the diagonal being ones. That is, there
are test items associated with each factor for sure and they may be associated
with other factors as well. Nonzero entries except for the diagonal entries of the
sub-matrix are penalized during the estimation procedure. Although, \code{"C2"}
is much weaker than \code{"C1"}, it still ensures empirical identifiability. Default is \code{"C1"}}

\item{non_pen}{the index of an item which is associated with each factor to satisfy \code{"C2"}.
For \code{C1}, the input can be \code{NULL}}

\item{gamma}{a numerical value of adaptive lasso parameter. Zou (2006) recommended three values, 0.5, 1, and 2.
The default value is 2.}
}
\value{
a list containing the following objects:
  \item{ra}{item discrimination parameters, a \eqn{J \times K} \code{matrix}}
  \item{rb}{item difficulty parameters, vector of length \eqn{J}}
  \item{reta}{variational parameters \eqn{\eta(\xi)}, a \eqn{N \times J} matrix}
  \item{reps}{variational parameters \eqn{\xi}, a \eqn{N \times J} matrix}
  \item{rsigma}{population variance-covariance matrix, a \eqn{K \times K} matrix}
  \item{mu_i}{mean parameter for each person, a \eqn{K \times N} matrix}
  \item{sig_i}{covariance matrix for each person, a \eqn{K \times K \times N} array}
  \item{n}{the number of iterations for the EM cycle}
  \item{Q_mat}{factor loading structure, a \eqn{J \times K} matrix}
  \item{GIC}{model fit index}
  \item{AIC}{model fit index}
  \item{BIC}{model fit index}
  \item{lbd}{numerical value of lasso penalty parameter \eqn{\lambda}}
}
\description{
Exploratory M2PL Analysis with Adaptive Lasso Penalty
}
\examples{
gvem_2PLEFA_adaptlasso(exampleData_2pl, exampleIndic_efa2pl_c1,constrain="C1",non_pen=NULL,gamma=2)
gvem_2PLEFA_adaptlasso(exampleData_2pl, exampleIndic_efa2pl_c2,constrain="C2",non_pen=61,gamma=2)
}
\references{
Cho, A. E., Xiao, J., Wang, C., & Xu, G. (2022). Regularized Variational Estimation for Exploratory Item Factor Analysis. \emph{Psychometrika}. https://doi.org/10.1007/s11336-022-09874-6

Zou, H. (2006). The adaptive LASSO and its oracle properties.  \emph{Journal of the American Statistical Association, 7}, 1011418–1429.
}
\seealso{
\code{\link{gvem_2PLEFA_rot}}, \code{\link{gvem_2PLEFA_lasso}}, \code{\link{exampleIndic_efa2pl_c1}}, \code{\link{exampleIndic_efa2pl_c2}}
}
