% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvem_2PLEFA_rot.R
\name{gvem_2PLEFA_rot}
\alias{gvem_2PLEFA_rot}
\title{Exploratory M2PL Analysis with Post-hoc Rotation}
\usage{
gvem_2PLEFA_rot(u, domain, max.iter = 5000, rot = "Promax")
}
\arguments{
\item{u}{a \eqn{N \times J} \code{matrix} or a \code{data.frame} that
consists of binary responses of \eqn{N} individuals to \eqn{J} items. The
missing values are coded as \code{NA}}

\item{domain}{the number of factors}

\item{max.iter}{the maximum number of iterations for the EM cycle; default is 5000}

\item{rot}{the post-hoc rotation method: Promax or CF-Quartimax; default is \code{"Promax"},
but may also be \code{"cfQ"} for conducting the CF-Quartimax rotation}
}
\value{
a list containing the following objects:
  \item{ra}{item discrimination parameters, a \eqn{J \times K} \code{matrix}}
  \item{rb}{item difficulty parameters, vector of length \eqn{J}}
  \item{reta}{variational parameters \eqn{\eta(\xi)}, a \eqn{N \times J} matrix}
  \item{reps}{variational parameters \eqn{\xi}, a \eqn{N \times J} matrix}
  \item{rsigma}{population variance-covariance matrix, a \eqn{K \times K} matrix}
  \item{mu_i}{mean parameter for each person, a \eqn{K \times N} matrix}
  \item{sig_i}{covariance matrix for each person, a \eqn{K \times K \times N} array}
  \item{n}{the number of iterations for the EM cycle}
  \item{rk}{factor loadings, a \eqn{J \times K} \code{matrix}}
  \item{Q_mat}{factor loading structure, a \eqn{J \times K} matrix}
  \item{GIC}{model fit index}
  \item{AIC}{model fit index}
  \item{BIC}{model fit index}
  \item{ur_a}{item discrimination parameters before conducting the rotation, a \eqn{J \times K} \code{matrix}}
}
\description{
Exploratory M2PL Analysis with Post-hoc Rotation
}
\examples{
\dontrun{
gvem_2PLEFA_rot(exampleData_2pl, domain=5,max.iter=3000)
gvem_2PLEFA_rot(exampleData_2pl, domain=5,rot="cfQ")
}
}
\seealso{
\code{\link{gvem_2PLEFA_lasso}}, \code{\link{gvem_2PLEFA_adaptlasso}}
}
