% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_emm.R
\name{lrt_DIF}
\alias{lrt_DIF}
\title{Likelihood Ratio Test for DIF Detection in 2PL Models}
\usage{
lrt_DIF(Y, D, X, unif = F)
}
\arguments{
\item{Y}{An \eqn{N\times J} binary matrix of item responses}

\item{D}{A \eqn{J\times K} binary matrix of loading indicators}

\item{X}{An \eqn{N} dimensional vector of group indicators (integers from \code{1} to \code{G})}

\item{unif}{Whether to detect uniform DIF only}
}
\value{
A list:
  \item{Sigma}{Group-level posterior covariance matrices}
  \item{Mu}{Group-level posterior mean vectors}
  \item{a}{Slopes for group 1}
  \item{b}{Intercepts for group 1}
  \item{gamma}{DIF parameters for the slopes}
  \item{beta}{DIF parameters for the intercepts}
}
\description{
Likelihood Ratio Test for DIF Detection in 2PL Models
}
\examples{
\dontrun{
with(exampleDIF, lrt_DIF(Y, D, X))}
}
\seealso{
\code{\link{em_DIF}}, \code{\link{gvemm_DIF}}, \code{\link{coef.vemirt_DIF}}, \code{\link{print.vemirt_DIF}}
}
