% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MGPCM_data}
\alias{MGPCM_data}
\title{Simulated Data Set for the Generalized Partial Credit Model}
\format{
An object of class \code{list} of length 3.
}
\usage{
MGPCM_data
}
\description{
Simulated Data Set for the Generalized Partial Credit Model
}
\keyword{datasets}
