% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPCM_gvem.R
\name{MGPCM_gvem}
\alias{MGPCM_gvem}
\title{GVEM Algorithm for the Generalized Partial Credit Model}
\usage{
MGPCM_gvem(
  data,
  model = matrix(1, nrow = J, ncol = 4),
  group = rep(1, nrow(data)),
  iter = 2000,
  eps = 1e-05,
  SE = FALSE,
  verbose = TRUE,
  EFA = FALSE
)
}
\arguments{
\item{data}{An \eqn{N\times J} matrix of item responses where 0 is the minimal partial credit score (missing responses should be coded as \code{NA})}

\item{model}{A \eqn{J\times K} matrix of loading indicators (K is the Number of latent dimension)(all items load on the only dimension by default)}

\item{iter}{Maximum number of iterations}

\item{eps}{Termination criterion on numerical accuracy}

\item{SE}{Whether to calculate the standard errors}

\item{verbose}{Whether to show the progress}

\item{EFA}{Whether to rotate the output}
}
\value{
An object of class \code{vemirt_DIF}, which is a list containing the following elements:
  \item{ ...$Sigma}{Group-level covariance matrices}#'   
  \item{ ...$MU}{Person-level posterior mean vectors}
  \item{ ...$a}{Slopes for group 1}
  \item{ ...$b}{Intercepts for group 1}
  \item{ ...$ll}{Estimated lower bound of log-likelihood}
}
\description{
GVEM Algorithm for the Generalized Partial Credit Model
}
\examples{
\dontrun{
with(MGPCM_data, MGPCM_gvem(data, model))}

}
\author{
Yijun Cheng <chengxb@uw.edu>
}
