% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MGRM_data}
\alias{MGRM_data}
\title{Simulated Data Set for the Graded Response Model}
\format{
A list of components of the data set:
\tabular{ll}{
\code{ data}\tab Item responses\cr
\tab\cr
\code{model}\tab Loading indicators\cr
\tab\cr
\code{params}\tab A list of true parameters used for generating the item responses:\cr
\tab\cr
\code{ ...$a}\tab Slopes\cr
\tab\cr
\code{ ...$b}\tab Negated intercepts\cr
\tab\cr
\code{ ...$theta}\tab Latent traits\cr
}
}
\usage{
MGRM_data
}
\description{
Simulated Data Set for the Graded Response Model
}
\author{
Yijun Cheng <chengxb@uw.edu>
}
\keyword{datasets}
