% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGRM_gvem.R
\name{MGRM_gvem}
\alias{MGRM_gvem}
\title{GVEM Algorithm for the Graded Response Model}
\usage{
MGRM_gvem(
  data,
  model = matrix(1, ncol(data)),
  method = "GVEM",
  iter = 200,
  tol = 1e-04,
  S = 10,
  M = 10,
  MinDim = 0,
  MaxDim = 0,
  verbose = FALSE,
  EFA = FALSE
)
}
\arguments{
\item{data}{An \eqn{N\times J} matrix of item responses where 0 is the minimal partial credit score (missing responses should be coded as \code{NA})}

\item{model}{A \eqn{J\times K} matrix of loading indicators (K is the Number of latent dimension)(all items load on the only dimension by default)}

\item{iter}{Maximum number of iterations}

\item{tol}{Termination criterion on numerical accuracy}

\item{S}{Sample size for approximating the expected lower bound (\code{'IWGVEM'} only)}

\item{M}{Sample size for approximating a tighter lower bound (\code{'IWGVEM'} only)}

\item{MinDim}{Minimum num of possible dimensions (\code{'EFA'} only)}

\item{MaxDim}{Maximum num of possible dimensions (\code{'EFA'} only)}

\item{verbose}{Whether to show the progress}

\item{EFA}{Whether to run EFA or CFA}

\item{criterion}{Information criterion for model selection, one of \code{'GIC'} (recommended), \code{'BIC'}, or \code{'AIC'}}

\item{c}{Constant for computing GIC}
}
\value{
An object of class \code{vemirt_DIF}, which is a list containing the following elements:
  \item{ ...$SIGMA}{Person-level posterior covariance matrices}
  \item{ ...$MU}{Person-level posterior mean vectors}
  \item{ ...$Sigma}{Group-level covariance matrices}
  \item{ ...$Mu}{Group-level mean vectors}
  \item{ ...$ksi1}{Variational parameter 1}
  \item{ ...$ksi2}{Variational parameter 2}
  \item{ ...$dim}{Num of dimension between latent variables}
  \item{ ...$a}{Slopes}
  \item{ ...$b}{Intercepts}
  \item{ ...$n2vlb}{Bayesian Information Criterion: \code{-2*ll+l0*log(N)}}
  \item{iter}{Number(s) of iterations for initialization}
}
\description{
GVEM Algorithm for the Graded Response Model
}
\examples{
\dontrun{
with(MGRM_data, MGRM_gvem(data, method = "IWGVEM", model, EFA = FALSE))}
}
\author{
Yijun Cheng <chengxb@uw.edu>
}
