% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_poly.R
\name{pa_poly}
\alias{pa_poly}
\title{Parallel analysis using polychoric correlation}
\usage{
pa_poly(data, n.iter = 10, figure = TRUE)
}
\arguments{
\item{data}{a \eqn{N \times J} \code{matrix} or a \code{data.frame} that
consists of the responses of \eqn{N} individuals to \eqn{J} items
without any missing values. The responses are binary or polytomous.}

\item{n.iter}{Number of simulated analyses to perform}

\item{figure}{By default, \code{pa_poly} draws an eigenvalue plot.
If FALSE, it suppresses the graphic output}
}
\value{
\code{pa_poly} returns a \code{data.frame} with the eigenvalues
for the real data and the simulated data.
}
\description{
Identify the number of factors
}
\examples{
\dontrun{
pa_poly(exampleData_2pl, n.iter=20)}
}
