% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.vemirt_DIF}
\alias{print.vemirt_DIF}
\title{Print DIF Items by Group}
\usage{
print(x, criterion = NULL, max = 99999L, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{vemirt_DIF}}

\item{criterion}{Information criterion for model selection, one of \code{'AIC'}, \code{'BIC'}, \code{'GIC'}, or the constant for computing GIC, otherwise use the criterion specified when fitting the model(s)}
}
\description{
Print DIF Items by Group
}
\seealso{
\code{\link{D1PL_em}}, \code{\link{D1PL_pair_em}}, \code{\link{D1PL_gvem}}, \code{\link{D2PL_em}}, \code{\link{D2PL_pair_em}}, \code{\link{D2PL_gvem}}, \code{\link{coef.vemirt_DIF}}, \code{\link{summary.vemirt_DIF}}
}
\author{
Weicong Lyu <weiconglyu@um.edu.mo>
}
