% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.vemirt_FA}
\alias{print.vemirt_FA}
\title{Print D2PL Items by Group}
\usage{
print(x)
}
\arguments{
\item{x}{An object of class \code{vemirt_FA}}
}
\description{
Print D2PL Items by Group
}
\seealso{
\code{\link{C2PL_gvem}}, \code{\link{C2PL_bs}}, \code{\link{C2PL_iw}}, \code{\link{C3PL_sgvem}}, \code{\link{E2PL_gvem_adaptlasso}}, \code{\link{E2PL_gvem_lasso}}, \code{\link{E2PL_gvem_rot}}, \code{\link{E2PL_IS}}, \code{\link{E3PL_sgvem_adaptlasso}}, \code{\link{E3PL_sgvem_lasso}}, \code{\link{E3PL_sgvem_rot}}, \code{\link{coef.vemirt_FA}}
}
\author{
Weicong Lyu <wlyu4@uw.edu>
}
