% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgvem_3PLCFA.R
\name{sgvem_3PLCFA}
\alias{sgvem_3PLCFA}
\title{Stochastic GVEM for Confirmatory M3PL Analysis}
\usage{
sgvem_3PLCFA(
  u,
  indic,
  samp = 50,
  forgetrate = 0.51,
  mu_b,
  sigma2_b,
  Alpha,
  Beta,
  max.iter = 5000
)
}
\arguments{
\item{u}{a \eqn{N \times J} \code{matrix} or a \code{data.frame} that
consists of binary responses of \eqn{N} individuals to \eqn{J} items. The
missing values are coded as \code{NA}}

\item{indic}{a \eqn{J \times K} \code{matrix} or a \code{data.frame} that
describes the factor loading structure of \eqn{J} items to \eqn{K} factors. It
consists of binary values where 0 refers to the item is irrelevant with this factor,
1 otherwise}

\item{samp}{a subsample for each iteration; default is 50}

\item{forgetrate}{the forget rate for the stochastic algorithm. The value should be within the
range from 0.5 to 1. Default is 0.51}

\item{mu_b}{the mean parameter for the prior distribution of item difficulty parameters}

\item{sigma2_b}{the variance parameter for the prior distribution of item difficulty parameters}

\item{Alpha}{the \eqn{\alpha} parameter for the prior distribution of guessing parameters}

\item{Beta}{the \eqn{\beta} parameter for the prior distribution of guessing parameters}

\item{max.iter}{the maximum number of iterations for the EM cycle; default is 5000}
}
\value{
a list containing the following objects:
  \item{ra}{item discrimination parameters, a \eqn{J \times K} \code{matrix}}
  \item{rb}{item difficulty parameters, vector of length \eqn{J}}
  \item{rc}{item guessing parameters, vector of length \eqn{J}}
  \item{rs}{variational parameters \eqn{s}, a \eqn{N \times J} matrix}
  \item{reta}{variational parameters \eqn{\eta(\xi)}, a \eqn{N \times J} matrix}
  \item{reps}{variational parameters \eqn{\xi}, a \eqn{N \times J} matrix}
  \item{rsigma}{population variance-covariance matrix, a \eqn{K \times K} matrix}
  \item{mu_i}{mean parameter for each person, a \eqn{K \times N} matrix}
  \item{sig_i}{covariance matrix for each person, a \eqn{K \times K \times N} array}
  \item{n}{the number of iterations for the EM cycle}
  \item{Q_mat}{factor loading structure, a \eqn{J \times K} matrix}
  \item{GIC}{model fit index}
  \item{AIC}{model fit index}
  \item{BIC}{model fit index}
}
\description{
Stochastic GVEM for Confirmatory M3PL Analysis
}
\examples{
\dontrun{
sgvem_3PLCFA(exampleData_3pl, exampleIndic_cfa3pl,samp=50,forgetrate=0.51,
mu_b=0,sigma2_b=4,Alpha=10,Beta=40)}
}
\references{
Cho, A. E., Wang, C., Zhang, X., & Xu, G. (2021). Gaussian variational estimation for multidimensional item response theory. \emph{British Journal of Mathematical and Statistical Psychology, 74}, 52-85.

Cho, A. E., Xiao, J., Wang, C., & Xu, G. (2022). Regularized Variational Estimation for Exploratory Item Factor Analysis. \emph{Psychometrika}. https://doi.org/10.1007/s11336-022-09874-6
}
\seealso{
\code{\link{gvem_2PLCFA}}
}
