% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgvem_3PLEFA_lasso.R
\name{sgvem_3PLEFA_lasso}
\alias{sgvem_3PLEFA_lasso}
\title{Stochastic GVEM with Lasso Penalty for Exploratory M3PL Analysis}
\usage{
sgvem_3PLEFA_lasso(
  u,
  indic,
  samp = 50,
  forgetrate = 0.51,
  mu_b,
  sigma2_b,
  Alpha,
  Beta,
  max.iter = 5000,
  constrain = "C1",
  non_pen = NULL
)
}
\arguments{
\item{u}{a \eqn{N \times J} \code{matrix} or a \code{data.frame} that
consists of binary responses of \eqn{N} individuals to \eqn{J} items. The
missing values are coded as \code{NA}}

\item{indic}{a \eqn{J \times K} \code{matrix} or a \code{data.frame} that
describes the factor loading structure of \eqn{J} items to \eqn{K} factors. It
consists of binary values where 0 refers to the item is irrelevant with this factor,
1 otherwise. For exploratory factor analysis with lasso penalty, \code{indic} should be
imposed certain constraints on the a \eqn{K \times K} sub-matrix to ensure identifiability.
The remaining parts do not assume any pre-specified zero structure but instead, the
appropriate lasso penalty would recover the true zero structure. Also see \code{constrain}}

\item{samp}{a subsample for each iteration; default is 50}

\item{forgetrate}{the forget rate for the stochastic algorithm. The value should be within the
range from 0.5 to 1. Default is 0.51}

\item{mu_b}{the mean parameter for the prior distribution of item difficulty parameters}

\item{sigma2_b}{the variance parameter for the prior distribution of item difficulty parameters}

\item{Alpha}{the \eqn{\alpha} parameter for the prior distribution of guessing parameters}

\item{Beta}{the \eqn{\beta} parameter for the prior distribution of guessing parameters}

\item{max.iter}{the maximum number of iterations for the EM cycle; default is 5000}

\item{constrain}{the constraint setting: \code{"C1"} or \code{"C2"}. To ensure
identifiablity, \code{"C1"} sets a \eqn{K \times K} sub-matrix of \code{indic} to be an
identity matrix.This constraint anchor \eqn{K} factors by designating \eqn{K} items that load solely on each factor respectively.
Note that the \eqn{K \times K} matrix does not have to appear at the top of the \code{indic} matrix.
\code{"C2"} sets the \eqn{K \times K} sub-matrix to be a lower triangular matrix with the diagonal being ones. That is, there
are test items associated with each factor for sure and they may be associated
with other factors as well. Nonzero entries (in the lower triangular part) except for the diagonal entries of the
sub-matrix are penalized during the estimation procedure. For instance, assume \eqn{K=3}, then the \code{"C2"} constraint will
imply the following submatrix: \eqn{C2=\begin{bmatrix} 1 & 0 & 0\\ 1 & 1 & 0\\ 1 & 1 & 1\\\end{bmatrix}}. As shown, item 1 is allowed to only
load on the first factor, item 2 will for sure load on the second factor but it may also load on the first factor (hence a penalty is added
on the \eqn{(2,1)} element of \code{"C1"}, i.e., \eqn{C2_{2,1}} ). Item 3 will for sure load on the third factor but it may also load on the
first two factors. However, note that for all remaining items their loading vector will all be \eqn{(1, 1, 1)} hence indistinguishable from the
third anchor item. Therefore, we need to alert the algorithm that this third anchor item will for sure load on the third factor, and
and whether or not it loads on the first two factors depends on the regularization results. Therefore, we need to specify
\code{"non_pen="} to identify the \eqn{K}th anchor item. Although, \code{"C2"} is much weaker than \code{"C1"}, it still ensures empirical identifiability. Default is \code{"C1"}.
During estimation, under both the \code{"C1"} and \code{"C1"} constraints, the population means and variances are constrained to be 0 and 1, respectively.}

\item{non_pen}{the index of an item which is associated with each factor to satisfy \code{"C2"}.
For \code{C1}, the input can be \code{NULL}}
}
\value{
a list containing the following objects:
  \item{ra}{item discrimination parameters, a \eqn{J \times K} \code{matrix}}
  \item{rb}{item difficulty parameters, vector of length \eqn{J}}
  \item{rc}{item guessing parameters, vector of length \eqn{J}}
  \item{rs}{variational parameters \eqn{s}, a \eqn{N \times J} matrix}
  \item{reta}{variational parameters \eqn{\eta(\xi)}, a \eqn{N \times J} matrix}
  \item{reps}{variational parameters \eqn{\xi}, a \eqn{N \times J} matrix}
  \item{rsigma}{population variance-covariance matrix, a \eqn{K \times K} matrix}
  \item{mu_i}{mean parameter for each person, a \eqn{K \times N} matrix}
  \item{sig_i}{covariance matrix for each person, a \eqn{K \times K \times N} array}
  \item{n}{the number of iterations for the EM cycle}
  \item{Q_mat}{factor loading structure, a \eqn{J \times K} matrix}
  \item{GIC}{model fit index}
  \item{AIC}{model fit index}
  \item{BIC}{model fit index}
  \item{lbd}{numerical value of lasso penalty parameter \eqn{\lambda}}
}
\description{
Stochastic GVEM with Lasso Penalty for Exploratory M3PL Analysis
}
\examples{
\dontrun{
sgvem_3PLEFA_lasso(exampleData_3pl, exampleIndic_efa3pl_c1,samp=50,
forgetrate=0.51,mu_b=0,sigma2_b=4,Alpha=10,Beta=40,max.iter=5000,
constrain="C1",non_pen=NULL)
sgvem_3PLEFA_lasso(exampleData_3pl, exampleIndic_efa3pl_c2,samp=50,
forgetrate=0.51,mu_b=0,sigma2_b=4,Alpha=10,Beta=40,max.iter=5000,
constrain="C2",non_pen=19)
}
}
\references{
Cho, A. E., Xiao, J., Wang, C., & Xu, G. (2022). Regularized Variational Estimation for Exploratory Item Factor Analysis. \emph{Psychometrika}. https://doi.org/10.1007/s11336-022-09874-6
}
\seealso{
\code{\link{sgvem_3PLEFA_rot}}, \code{\link{sgvem_3PLEFA_adaptlasso}}, \code{\link{exampleIndic_efa3pl_c1}}, \code{\link{exampleIndic_efa3pl_c2}}
}
