% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgvem_3PLEFA_rot.R
\name{sgvem_3PLEFA_rot}
\alias{sgvem_3PLEFA_rot}
\title{Stochastic GVEM for Exploratory M3PL Analysis}
\usage{
sgvem_3PLEFA_rot(
  u,
  domain,
  samp = 50,
  forgetrate = 0.51,
  mu_b,
  sigma2_b,
  Alpha,
  Beta,
  max.iter = 5000,
  rot = "Promax"
)
}
\arguments{
\item{u}{a \eqn{N \times J} \code{matrix} or a \code{data.frame} that
consists of binary responses of \eqn{N} individuals to \eqn{J} items. The
missing values are coded as \code{NA}}

\item{domain}{the number of factors}

\item{samp}{a subsample for each iteration; default is 50}

\item{forgetrate}{the forget rate for the stochastic algorithm. The value should be within the
range from 0.5 to 1. Default is 0.51}

\item{mu_b}{the mean parameter for the prior distribution of item difficulty parameters}

\item{sigma2_b}{the variance parameter for the prior distribution of item difficulty parameters}

\item{Alpha}{the \eqn{\alpha} parameter for the prior distribution of guessing parameters}

\item{Beta}{the \eqn{\beta} parameter for the prior distribution of guessing parameters}

\item{max.iter}{the maximum number of iterations for the EM cycle; default is 5000}

\item{rot}{the post-hoc rotation method: Promax or CF-Quartimax; default is \code{"Promax"},
but may also be \code{"cfQ"} for conducting the CF-Quartimax rotation}
}
\value{
a list containing the following objects:
  \item{ra}{item discrimination parameters, a \eqn{J \times K} \code{matrix}}
  \item{rb}{item difficulty parameters, vector of length \eqn{J}}
  \item{rc}{item guessing parameters, vector of length \eqn{J}}
  \item{rs}{variational parameters \eqn{s}, a \eqn{N \times J} matrix}
  \item{reta}{variational parameters \eqn{\eta(\xi)}, a \eqn{N \times J} matrix}
  \item{reps}{variational parameters \eqn{\xi}, a \eqn{N \times J} matrix}
  \item{rsigma}{population variance-covariance matrix, a \eqn{K \times K} matrix}
  \item{mu_i}{mean parameter for each person, a \eqn{K \times N} matrix}
  \item{sig_i}{covariance matrix for each person, a \eqn{K \times K \times N} array}
  \item{n}{the number of iterations for the EM cycle}
  \item{Q_mat}{factor loading structure, a \eqn{J \times K} matrix}
  \item{rk}{factor loadings, a \eqn{J \times K} \code{matrix}}
  \item{GIC}{model fit index}
  \item{AIC}{model fit index}
  \item{BIC}{model fit index}
  \item{ur_a}{item discrimination parameters before conducting the rotation, a \eqn{J \times K} \code{matrix}}
}
\description{
Stochastic GVEM for Exploratory M3PL Analysis
}
\examples{
\dontrun{
sgvem_3PLEFA_rot(exampleData_3pl, 3,samp=50,forgetrate=0.51,
mu_b=0,sigma2_b=4,Alpha=10,Beta=40,max.iter=5000,rot="Promax")}
}
\seealso{
\code{\link{sgvem_3PLEFA_lasso}}, \code{\link{sgvem_3PLEFA_adaptlasso}}
}
