% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vemirt_DIF.R
\name{coef.vemirt_DIF}
\alias{coef.vemirt_DIF}
\title{Extract Parameter Estimates from DIF Analysis}
\usage{
coef(object, criterion = NULL)
}
\arguments{
\item{object}{An object of class \code{vemirt_DIF}}

\item{criterion}{Information criterion for model selection, one of \code{'AIC'}, \code{'BIC'} and \code{'GIC'}, otherwise use the criterion specified when fitting the model(s)}
}
\description{
Extract Parameter Estimates from DIF Analysis
}
\seealso{
\code{\link{em_DIF}}, \code{\link{gvemm_DIF}}, \code{\link{lrt_DIF}}, \code{\link{print.vemirt_DIF}}
}
