% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_em.R
\name{DIF_em}
\alias{DIF_em}
\title{EM Algorithms for DIF Detection in 2PL Models}
\usage{
DIF_em(
  data,
  model = matrix(1, ncol(data)),
  group = rep(1, nrow(data)),
  method = "EMM",
  Lambda0 = seq(0.1, 0.8, by = 0.1),
  level = 12,
  criterion = "BIC",
  iter = 200,
  eps = 0.001,
  lr = 1,
  c = 1
)
}
\arguments{
\item{data}{An \eqn{N\times J} binary matrix of item responses}

\item{model}{A \eqn{J\times K} binary matrix of loading indicators  (all items load on the only dimension by default)}

\item{group}{An \eqn{N} dimensional vector of group indicators from \code{1} to \code{G} (all respondents are in the same group by default)}

\item{method}{Estimation algorithm, one of \code{'EM'} or \code{'EMM'}}

\item{Lambda0}{A vector of \code{lambda0} values for \eqn{L_1} penalty (\code{lambda} equals \code{sqrt(N) * lambda0})}

\item{level}{Accuracy level, either a number for \code{mvQuad} or a vector indicating the grid for each latent dimension}

\item{criterion}{Information criterion for model selection, one of \code{'BIC'} (recommended), \code{'AIC'}, or \code{'GIC'}}

\item{iter}{Maximum number of iterations}

\item{eps}{Termination criterion on numerical accuracy}

\item{lr}{Learning rate for the Newton-Raphson method}

\item{c}{Constant for computing GIC}
}
\value{
An object of class \code{vemirt_DIF}, which is a list containing the following elements:
  \item{N}{Number of respondents}
  \item{niter0}{Number(s) of iterations for initialization}
  \item{fit}{The best (with lowest information criterion) model, which is an element of \code{all}}
  \item{best}{The index of \code{fit} in \code{all}}
  \item{all}{A list of models which has the same length as \code{Lambda0}:}
  \item{ ...$lambda0}{Corresponding element in \code{Lambda0}}
  \item{ ...$lambda}{\code{sqrt(N) * lambda0}}
  \item{ ...$niter}{Number(s) of iterations}
  \item{ ...$Sigma}{Group-level posterior covariance matrices}
  \item{ ...$Mu}{Group-level posterior mean vectors}
  \item{ ...$a}{Slopes for group 1}
  \item{ ...$b}{Intercepts for group 1}
  \item{ ...$gamma}{DIF parameters for the slopes}
  \item{ ...$beta}{DIF parameters for the intercepts}
  \item{ ...$ll}{Log-likelihood}
  \item{ ...$l0}{Number of nonzero DIF parameters in \code{gamma} and \code{beta}}
  \item{ ...$AIC}{Akaike Information Criterion: \code{-2*ll+l0*2}}
  \item{ ...$BIC}{Bayesian Information Criterion: \code{-2*ll+l0*log(N)}}
  \item{ ...$GIC}{Generalized Information Criterion: \code{-2*ll+c*l0*log(N)*log(log(N))}}
}
\description{
EM Algorithms for DIF Detection in 2PL Models
}
\examples{
\dontrun{
with(exampleDIF, DIF_em(data, model, group))}
}
\seealso{
\code{\link{DIF_gvem}}, \code{\link{DIF_lrt}}, \code{\link{coef.vemirt_DIF}}, \code{\link{print.vemirt_DIF}}, \code{\link{summary.vemirt_DIF}}
}
\author{
Weicong Lyu <wlyu4@uw.edu>
}
