% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MGRM_data}
\alias{MGRM_data}
\title{Simulated Data Set for the Graded Response Model}
\format{
An object of class \code{list} of length 3.
}
\usage{
MGRM_data
}
\description{
Simulated Data Set for the Graded Response Model
}
\keyword{datasets}
