% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vemirt_DIF.R
\name{print.vemirt_DIF}
\alias{print.vemirt_DIF}
\title{Print DIF Items}
\usage{
print(x, criterion = NULL)
}
\arguments{
\item{x}{An object of class \code{vemirt_DIF}}

\item{criterion}{Information criterion for model selection, one of \code{'AIC'}, \code{'BIC'} and \code{'GIC'}, otherwise use the criterion specified when fitting the model(s)}
}
\description{
Print DIF Items
}
\seealso{
\code{\link{em_DIF}}, \code{\link{gvemm_DIF}}, \code{\link{lrt_DIF}}, \code{\link{coef.vemirt_DIF}}
}
