% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.vemirt_DIF_summary}
\alias{print.vemirt_DIF_summary}
\title{Print Summary of DIF 2PL Items}
\usage{
print(x, max = 99999L, ...)
}
\arguments{
\item{x}{An object of class \code{vemirt_DIF_summary}}
}
\description{
Print Summary of DIF 2PL Items
}
\seealso{
\code{\link{summary.vemirt_DIF}}, \code{\link{coef.vemirt_DIF_summary}}
}
\author{
Weicong Lyu <wlyu4@uw.edu>
}
