% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_utility.R
\name{summary.vemirt_DIF}
\alias{summary.vemirt_DIF}
\title{Summarize DIF Items}
\usage{
print(x, criterion = NULL)
}
\arguments{
\item{x}{An object of class \code{vemirt_DIF}}

\item{criterion}{Information criterion for model selection, one of \code{'AIC'}, \code{'BIC'}, \code{'GIC'}, or the constant for computing GIC, otherwise use the criterion specified when fitting the model(s)}
}
\description{
Summarize DIF Items
}
\seealso{
\code{\link{em_DIF}}, \code{\link{gvemm_DIF}}, \code{\link{lrt_DIF}}, \code{\link{coef.vemirt_DIF}}, \code{\link{print.vemirt_DIF}}
}
