% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMIRT-package.R
\docType{package}
\name{VEMIRT-package}
\alias{VEMIRT}
\alias{VEMIRT-package}
\title{VEMIRT: A Package for High-Dimensional IRT Models}
\description{
VEMIRT is created to assist researchers to conduct exploratory and confirmatory
multidimensional item response theory (MIRT) analysis and cooresponding item
differential functioning (DIF) analysis. The core computation
engine of VEMIRT is a family of Gaussian Variational EM algorithms that are
considerably more efficient than currently available algorithms in other
software packages, especially when the number of latent factors exceeds four.
}
\section{Identifying the number of factors}{

\code{\link{pa_poly}} identifies the number of factors via parallel analysis.
}

\section{Exploratory factor analysis}{

\itemize{
  \item \code{\link{E2PL_gvem_rot}} conducts M2PL Analysis with post-hoc rotation (Promax & CF-Quartimax)
  \item \code{\link{E2PL_gvem_lasso}} conducts M2PL Analysis with Lasso penalty
  \item \code{\link{E2PL_gvem_adaptlasso}} conducts M2PL Analysis with adaptive Lasso penalty
  \item \code{\link{E2PL_iw}} conducts importance sampling to correct bias for M2PL analysis
  \item \code{\link{E3PL_sgvem_rot}} conducts stochastic GVEM to further improve the computational effficiency for exploratory M3PL analysis
  \item \code{\link{E3PL_sgvem_lasso}} conducts M3PL Analysis with Lasso penalty
  \item \code{\link{E3PL_sgvem_adaptlasso}} conducts M3PL Analysis with adaptive Lasso penalty
  \item \code{\link{MGRM_gvem}} conducts GVEM for the multidimensional graded response model with post-hoc rotation
  \item \code{\link{MGPCM_gvem}} conducts GVEM for the multidimensional partial credit model with post-hoc rotation
}
}

\section{Confirmatory factor analysis}{

\itemize{
  \item \code{\link{C2PL_gvem}} conducts GVEM for confirmatory M2PL analysis
  \item \code{\link{C2PL_bs}} conducts bootstrap sampling to correct bias and produce standard errors for confirmatory M2PL analysis
  \item \code{\link{C2PL_iw}} conducts importance sampling to correct bias for M2PL analysis
  \item \code{\link{C2PL_iw2}} conducts IW-GVEM for confirmatory M2PL analysis (alternative implementation to \code{\link{C2PL_iw}})
  \item \code{\link{C3PL_sgvem}} conducts stochastic GVEM for confirmatory M3PL analysis
  \item \code{\link{MGRM_gvem}} conducts GVEM for the multidimensional graded response model
  \item \code{\link{MGPCM_gvem}} conducts GVEM for the multidimensional partial credit model
}
}

\section{Differential item functioning analysis}{

\itemize{
  \item \code{\link{D1PL_em}} conducts DIF analysis for M1PL models using EM algorithms
  \item \code{\link{D1PL_gvem}} conducts DIF analysis for M1PL models using GVEM algorithms
  \item \code{\link{D1PL_pair_em}} conducts DIF analysis for 1PL models using EM algorithms with group pairwise truncated \eqn{L_1} penalty
  \item \code{\link{D2PL_em}} conducts DIF analysis for M2PL models using EM algorithms
  \item \code{\link{D2PL_pair_em}} conducts DIF analysis for 2PL models using EM algorithms with group pairwise truncated \eqn{L_1} penalty
  \item \code{\link{D2PL_gvem}} conducts DIF analysis for M2PL models using GVEM algorithms
  \item \code{\link{D2PL_lrt}} conducts DIF analysis for M2PL models using the likelihood ratio test
  \item \code{\link{D2PL_bifactor_gvem}} conducts DIF analysis for M2PL bifactor models using GVEM algorithms
}
}

\section{Shiny apps for VEMIRT}{

\itemize{
  \item \code{\link{shinyVEMIRT}} Run the shiny app for VEMIRT
  \item \code{\link{DIFdashboard}} Run the shiny app for DIF Dashboard
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://MAP-LAB-UW.github.io/VEMIRT}
  \item \url{https://github.com/MAP-LAB-UW/VEMIRT}
}

}
\author{
\strong{Maintainer}: Weicong Lyu \email{weiconglyu@um.edu.mo} (\href{https://orcid.org/0000-0002-8159-2161}{ORCID})

Authors:
\itemize{
  \item Yijun Cheng \email{chengxb@uw.edu} (\href{https://orcid.org/0000-0002-0671-9193}{ORCID})
  \item Jiaying Xiao \email{jxiao6@uw.edu} (\href{https://orcid.org/0000-0001-9513-6477}{ORCID})
  \item He Ren \email{heren@uw.edu} (\href{https://orcid.org/0000-0003-1208-7089}{ORCID})
  \item Ruoyi Zhu \email{zhux0445@uw.edu} (\href{https://orcid.org/0000-0002-8159-2161}{ORCID})
  \item Gongjun Xu \email{gongjun@umich.edu} (\href{https://orcid.org/0000-0003-4023-5413}{ORCID})
  \item Chun Wang \email{wang4066@uw.edu} (\href{https://orcid.org/0000-0003-2695-9781}{ORCID})
}

}
