% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMIRT-package.R
\docType{package}
\name{VEMIRT-package}
\alias{VEMIRT}
\alias{VEMIRT-package}
\title{VEMIRT: A package for high-dimensional IRT models}
\description{
VEMIRT is created to assist researchers to conduct exploratory and confirmatory
multidimensional item response theory (MIRT) analysis and cooresponding item
differential functioning (DIF) analysis. The core computation
engine of VEMIRT is a family of Gaussian Variational EM algorithms that are
considerably more efficient than currently available algorithms in other
software packages, especially when the number of latent factors exceeds four.
}
\section{Identifying the number of factors}{

\code{\link{pa_poly}} identifies the number of factors via parallel analysis.
}

\section{Exploratory factor analysis}{

\itemize{
  \item \code{\link{gvem_2PLEFA_rot}} conducts M2PL Analysis with post-hoc rotation (Promax & CF-Quartimax)
  \item \code{\link{gvem_2PLEFA_lasso}} conducts M2PL Analysis with Lasso penalty
  \item \code{\link{gvem_2PLEFA_adaptlasso}} conducts M2PL Analysis with adaptive Lasso penalty
  \item \code{\link{sgvem_3PLEFA_rot}} conducts stochastic GVEM to futher imporve the computational effficiency for exploratory M3PL analysis
  \item \code{\link{sgvem_3PLEFA_lasso}} conducts M3PL Analysis with Lasso penalty
  \item \code{\link{sgvem_3PLEFA_adaptlasso}} conducts M3PL Analysis with adaptive Lasso penalty
}
}

\section{Confirmatory factor analysis}{

\itemize{
  \item \code{\link{gvem_2PLCFA}} conducts GVEM for confirmatory M2PL analysis
  \item \code{\link{sgvem_3PLCFA}} conducts stochastic GVEM for confirmatory M3PL analysis
  \item \code{\link{bs_2PLCFA}} conducts bootstrap sampling to correct bias and produce standard errors for confirmatory M2PL analysis
  \item \code{\link{importanceSampling}} conducts importance sampling to correct bias for M2PL analysis
}
}

\section{Differential item functioning analysis}{

\itemize{
  \item \code{\link{em_DIF}} conducts DIF analysis for M2PL models using EM algorithms
  \item \code{\link{gvemm_DIF}} conducts DIF analysis for M2PL models using GVEMM algorithms
  \item \code{\link{lrt_DIF}} conducts DIF analysis for M2PL models using the likelihood ratio test
}
}

\author{
\strong{Maintainer}: Jiaying Xiao \email{jxiao6@uw.edu} (\href{https://orcid.org/0000-0001-9513-6477}{ORCID})

Authors:
\itemize{
  \item Weicong Lyu \email{wlyu4@uw.edu} (\href{https://orcid.org/0000-0002-8159-2161}{ORCID})
  \item Ruoyi Zhu \email{zhux0445@uw.edu} (\href{https://orcid.org/0000-0002-8159-2161}{ORCID})
  \item Gongjun Xu \email{gongjun@umich.edu} (\href{https://orcid.org/0000-0003-4023-5413}{ORCID})
  \item Chun Wang \email{wang4066@uw.edu} (\href{https://orcid.org/0000-0003-2695-9781}{ORCID})
}

}
